Updated assign3.cpp
assign3.cpp now compiles and creates an executable without any errors but there are still some problems. choosing option 1 and inputing 'r' as the book type will cause the program to get stuck in an infinite loop. And you also have to work on the restrictors because if I try jumping to page 2000 it says "you jumped forward by 2000" but I stopped jumping over the max or below the min number of pages in book.cpp so it does not match. You should do the same thing for jumping backward. It should tell them that you cannot jump that far forward or backward or something.